XWIKI-19070: Change the notifications locations inclusion default: don't consider that the whole wiki is watched when nothing is watched
This commits provide the modification that have been voted to change the
standard behaviour of watched pages for new users.

The following behaviour are changed:
  1. by default, nothing is watched until the user decided to watch
     something
  2. by default, all events type are enabled until an admin explicitely
     switched off some event types, or the user does
  3. directly following 1 and 2, no event type is magically enabled anymore
     whenever a user creates a page with autowatch enabled

Technically (3) means that we don't need anymore XWikiEventTypesEnabler
component so I get rid of it. (1) means that the default value in
DefaultWatchedEntitiesManager needed to be changed as well as the way to
compute it in the script service: no preferences means no watch.
(2) is the one involving most of the changes: I needed to ensure that
default value for preferences is enabled, and to also ensure that
whenever the list of preferences is empty it means that everything is
enabled, which impacts directly UserEventManager to dispatch the events.
(2) also means that we don't need the
NotificationAdministrationDocumentInitializer for Mention anymore, as by
default this event type is now enabled, as well as all the other ones.

I updated the integration tests according to the changes.

Note that those changes don't involve any migration: existing users that
already have preferences should still use their preferences. The
behavioural change only applies to new users without any preferences
(including without any preferences set at administration level).